server/requirements.txt
python-docx
comtypes
djangorestframework
django-cors-headers